GamesHub/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
attrs==25.4.0
billiard==4.2.2
boto3==1.40.59
botocore==1.40.59
celery==5.5.3
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.4
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==45.0.4
deprecation==2.1.0
dj-database-url==3.0.1
Django==5.2.3
django-celery-beat==2.8.1
django-redis==6.0.0
django-storages==1.14.6
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
dotenv==0.9.9
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
et_xmlfile==2.0.0
eventlet==0.40.3
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
multidict==6.7.0
numpy==2.3.3
openpyxl==3.1.5
oracledb==3.1.1
packaging==25.0
pandas==2.3.2
pillow==12.0.0
pip==25.1.1
postgrest==2.24.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2==2.9.11
pycparser==2.22
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
realtime==2.24.0
redis==7.0.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
whitenoise==6.11.0
yarl==1.22.0
